Created project for stage 2.
I also used the assembler from stage 1 to compile all assembly files that do not contain symbols. Tested the files in the CPU Simulator, and they work.

I had to make some slight modifications to the assembler to flag when there is an error in the code. I had to change the instructions so that for C Instructions, the destinations can go in any order. There was one line in the pong game that had "MD=" instead of "DM=".

Translation for the destination part was modified to work Without Loss Of Generality. Improvements could be done for the assembler to stop when it encounters an error (and also to recognize errors).